refactor(form): use TimeInput primitive directly; drop time-picker changes
Form recipes were importing TimePickerFieldInput from the time-picker
demo file, which had to add ~150 lines (helpers + a new exported
component) to expose form-friendly plumbing. The Figma layout doesn't
use the picker popover for time — it shows a plain TimeInput. Switching
to TimeInput directly removes the need to touch time-picker.tsx at all,
and the form file only gains two small string<->{hour, minute}
helpers.

- src/app/demo/[name]/ui/time-picker.tsx: fully reverted to origin/main
- src/app/demo/[name]/ui/form-examples.tsx (was form-interactive.tsx):
  use @/components/ui/time-input's TimeInput directly; add inline
  parseTimeFieldString / formatTimeFieldValue helpers
- src/app/demo/[name]/ui/form.tsx: import path follows the rename
- registry.json: form example file path follows the rename
- src/app/(registry)/registry/[name]/page.tsx: example-code lookup
  follows the rename
